Finished implementation of user authentication
